audit-patch redeploy: MerklePriceOracle (Base 0xE49104b3 + LiteForge 0x20A81230) and GradedPriceOracle (LiteForge 0x6cca6D77)
Ships the de2b673 audit patches the original deploys predated: a
renounceOwnership lockout guard (a renounce on a write-once oracle
permanently bricks root updates) and uint256 categoryId on merkle. The type
widening is a validation fix only - abi.encode pads uint16 and uint256
identically, so leaves, roots and existing proofs are unchanged (proven:
rebuilt tree root == cached root; live proof verified end-to-end on the new
Base contract). Guard proven on all three via eth_call revert. Current root
pushed to all three; deployment JSONs superseded with the old records kept;
ABIs rewritten; updater comment documents the encoding invariant. Base
source verified on Basescan same-session.

+++ merkle_root_updater.py
@@ -33,8 +33,12 @@ def keccak256(data: bytes) -> bytes:
 
 
 def compute_leaf(pid, cat, name, market, low):
+    # uint256 categoryId as of the 2026-07-31 audit-patch redeploy (de2b673
+    # widened it from uint16). MUST match the deployed contract's computeLeaf
+    # abi.encode exactly or no proof will ever verify — if you repoint this
+    # updater at a different contract generation, check that signature first.
     inner = abi_encode(
-        ["uint256", "uint16", "string", "uint256", "uint256"],
+        ["uint256", "uint256", "string", "uint256", "uint256"],
         [pid, cat, name, market, low]
     )
     return keccak256(keccak256(inner))
